Refactor into separate modules; improve model relationships
