+++ collection/prefect_intel/packaging/__init__.py
@@ -1,0 +1,1 @@
+from .execution import run_in_environment
